2015-05-18 11PM
此次提交：
1. 修复 后台用户管理模块 在删除用户时 只删除数据库资料 而不删除 用户 头像文件的BUG
2. 修复后台用户管理模块，当删除一个用户的时候，同时也删除该用户上传的文章、相册、资源。
3. 后台资源链接管理在添加资源链接的时候的重复添加问题
4. 后台资源链接管理 在添加资源链接的增加 uid 的字段
5. 修复后台用户管理模块当无用户时的错误显示

+++ test/libs/Model/backUser/userModel.class.php
@@ -2,7 +2,11 @@
 
 class userModel{
 
-	public $_tableName2 = 'user_info';
+	private $_tableName1 = 'article_info';
+	private $_tableName2 = 'user_info';
+	private $_tableName3 = 'album_cover';
+	private $_tableName4 = 'photo_content';
+	private $_tableName5 = 'resource_info';
 
 
 //  用户列表页的查询操作
@@ -20,12 +24,12 @@ class userModel{
 		$sql = "SELECT * FROM $this->_tableName2 WHERE id=$id";
 		$res = DB::findAll($sql);
 		return $res;
-
+ 
 	}
 
 
 //  用户信息修改操作
-	public function userInfoChangeWork($POST,$GET){
+	public function userInfoChangeWork( $POST , $GET ){
 
 //  将生日拼起来,例如:1997-01-09
 		$birthday = $POST['year'].'-'.$POST['month'].'-'.$POST['day']; 
@@ -64,15 +68,100 @@ class userModel{
 
 //  用户删除操作
 	public function userDel($id){
-		
-		$sql = "DELETE FROM $this->_tableName2  WHERE id=$id";
-		$res = DB::query($sql);
-		return $res;
+		//$this->userDelFolder( $id ); exit; 
+		     
+
+		$this->userDelArticle( $id );  // 用户删除连带 删除该用户发表的文章
+		$this->userDelAlbum( $id );    // 用户删除连带 删除该用户发表的相册
+		$this->userDelResource( $id ); // 用户删除连带 删除该用户发表的资源
+
+
+
+	 	if( $this->userDelFolder( $id ) ){
+
+			$sql = "DELETE FROM $this->_tableName2  WHERE id=$id";
+			$DbDelRes = DB::query($sql);
+
+			if( $DbDelRes ){
+				return true; // 删除成功
+			}else{
+				return false;
+			}
+
+		}else{
+			return false;
+		}
 
 	}
 
 
 
+
+/*  
+	调用处: userDel()
+	作用: 用户删除连带 删除该用户发表的文章
+*/
+	public function userDelArticle( $uid ){
+
+		$sql = "DELETE FROM $this->_tableName1 WHERE uid=$uid";
+		$res = DB::query($sql);
+		return $res; //  返回 t / f
+		
+	}
+
+/*  
+	调用处：userDel()
+	作用: 用户删除连带 删除该用户发表的相册
+*/
+	public function userDelAlbum( $uid ){
+
+		$sql = "SELECT id FROM $this->_tableName3 WHERE uid=$uid";
+
+		$res   = DB::findAll( $sql );
+		$count = count( $res );
+
+		if( empty($res) ){ return false; }  // 如果删除用户时, 该用户并没有发布过相册的话, 就不用执行下面
+
+		for( $i=0; $i<$count; $i++ ){
+
+			$albumID = $res[$i]['id'];
+			//  从其他模型里,调用删除相册的功能模块
+			M('album','backPhoto')->albumDel( $albumID );  // 传入当前要删除的相册ID 
+		}
+	}
+
+/*  
+	调用处: userDel();
+	作用: 用户删除连带 删除该用户发表的资源
+*/
+	public function userDelResource( $uid ){
+		$sql = "DELETE FROM $this->_tableName5 WHERE uid=$uid";
+		$res = DB::query( $sql );
+	}
+
+
+
+
+/* 
+	调用处: 本类中: userDel();
+	作用：根据 id 删除该用户保存在 ‘头像文件夹’  中的头像文件
+*/
+	public function userDelFolder( $id ){
+		$sql = "SELECT photo FROM $this->_tableName2 WHERE id=$id";
+		$res = DB::findOne( $sql );
+
+		if( $res['photo'] == 'pictureGroup/userPhotoFolder/defaultPhoto.jpg' ){
+			// 如果当前要删除的用户 的头像时默认头像的话, 那就不用删除这个默认的头像文件
+			return true;
+		}
+		
+		if( unlink( $res['photo'] ) ){ // 删除该用户的 头像文件
+			return true;
+		}else{
+			return false;
+		}
+
+	}
 
 }
